+++ languages/c/templates/declarations/polymorphic-pull-event.c
@@ -2,3 +2,4 @@
 typedef void* (*${info.Title}${method.Name}Callback)( const void* userData, ${method.pulls.param.type} );
 int32_t ${info.Title}_Register_${method.Name}( ${info.Title}${method.Name}Callback userCB, const void* userData );
 int32_t ${info.Title}_Unregister_${method.Name}( ${info.Title}${method.Name}Callback userCB);
+