+++ languages/c/templates/declarations/polymorphic-pull.c
@@ -1,2 +1,3 @@
 /* ${method.name} - ${method.description} */
 int32_t ${info.Title}_Push${method.Name}( ${method.signature.params} );
+