+++ languages/c/templates/declarations/property.c
@@ -2,4 +2,4 @@
  * ${method.summary}
  * ${method.params}
  */
-int F${info.title}_Get${method.Name}(${method.signature.params}${if.params}, ${end.if.params}${method.result.properties});
+${method.signature}