+++ languages/c/templates/methods/default.c
@@ -1,5 +1,5 @@
 /* ${method.rpc.name} - ${method.description} */
-int32_t ${info.Title}_${method.Name}( ${method.signature.params}${if.result}${if.params}, ${end.if.params} ${method.result.type}* ${method.result.name}${end.if.result}${if.signature.empty}void${end.if.signature.empty} ) {
+int32_t ${info.Title}_${method.Name}( ${method.signature.params}${if.result}${if.params}, ${end.if.params}OUT ${method.result.type}* ${method.result.name}${end.if.result}${if.signature.empty}void${end.if.signature.empty} ) {
 
     int32_t status = FireboltSDKErrorUnavailable;
     FireboltSDK::Transport<WPEFramework::Core::JSON::IElement>* transport = FireboltSDK::Accessor::Instance().GetTransport();
@@ -19,3 +19,4 @@ int32_t ${info.Title}_${method.Name}( ${method.signature.params}${if.result}${if
   
     return status;
 }
+