+++ languages/c/templates/methods/polymorphic-pull.c
@@ -20,3 +20,4 @@ int32_t ${info.Title}_Push${method.Name}( ${method.signature.params} )
 
     return status;
 }
+