+++ languages/c/templates/schemas/src/jsondata_module.h
@@ -25,6 +25,6 @@ ${if.schemas}namespace FireboltSDK {
     namespace ${info.Title} {
     // Types
 
-       /* ${SCHEMAS:json-types} */
+       /* ${SCHEMAS} */
     }
 }${end.if.schemas}