+++ src/shared/modules.mjs
@@ -96,13 +96,13 @@ function getProviderInterface(capability, module, extractProviderSchema = false)
       // remove `onRequest`
       method.name = method.name.charAt(9).toLowerCase() + method.name.substr(10)
 
-        method.params = [
-          {
-            "name": "parameters",
-            "required": true,
-            "schema": payload.properties.parameters
-          }
-        ]
+      method.params = [
+        {
+          "name": "parameters",
+          "required": true,
+          "schema": payload.properties.parameters
+        }
+      ]
   
       if (!extractProviderSchema) {
         let exampleResult = null
@@ -619,21 +619,16 @@ const createResponseFromProvider = (provider, type, json) => {
     if (provider.tags.find(t => t[`x-${type.toLowerCase()}`])) {
         response.params = [
             {
-                name: type.toLowerCase(),
-                required: true,
+                name: "correlationId",
                 schema: {
-                    allOf: [
-                        {
-                            "$ref": "https://meta.comcast.com/firebolt/types#/definitions/ProviderResponse" // use this schema for both Errors and Results
-                        },
-                        {
-                            "type": "object",
-                            "properties": {
-                                "result": provider.tags.find(t => t[`x-${type.toLowerCase()}`])[`x-${type.toLowerCase()}`]
-                            }
-                        }
-                    ]
-                }
+                    type: "string"
+                },
+                required: true
+            },
+            {
+                name: type === 'Error' ? 'error' : "result",
+                schema: provider.tags.find(t => t[`x-${type.toLowerCase()}`])[`x-${type.toLowerCase()}`],
+                required: true
             }
         ]
 
@@ -674,11 +669,12 @@ const createResponseFromProvider = (provider, type, json) => {
                 name: schema.examples.length === 1 ? "Example" : `Example #${n++}`,
                 params: [
                     {
-                        name: `${type.toLowerCase()}`,
-                        value: {
-                            correlationId: "123",
-                            result: param
-                        }
+                        name: 'correlationId',
+                        value: '123'
+                    },
+                    {
+                        name: 'result',
+                        value: param
                     }
                 ],
                 result: {
@@ -717,11 +713,12 @@ const createResponseFromProvider = (provider, type, json) => {
                 name: 'Example 1',
                 params: [
                     {
-                        name: `${type.toLowerCase()}`,
-                        value: {
-                            correlationId: "123",
-                            result: value
-                        }                        
+                        name: 'correlationId',
+                        value: '123'
+                    },
+                    {
+                        name: type === 'Error' ? 'error' : 'result',
+                        value
                     }
                 ],
                 result: {